fix temp dtype

+++ verl/workers/engine/fsdp/transformer_impl.py
@@ -734,7 +734,7 @@ class FSDPEngineWithLMHead(FSDPEngine):
         position_ids = micro_batch["position_ids"]
 
         if not isinstance(temperature, torch.Tensor):
-            temperature = torch.tensor([temperature] * input_ids.shape[0], device=input_ids.device)
+            temperature = torch.tensor([temperature] * input_ids.shape[0], device=input_ids.device, dtype=torch.float32)
 
         assert temperature.shape[0] == input_ids.shape[0]
